+++ amplience/client.go
@@ -77,13 +77,35 @@ func ParseAndUnmarshalAmplienceResponseBody(response *http.Response, data interf
 
 // https://amplience.com/docs/api/dynamic-content/management/index.html#section/Usage/Status-Code-Table
 func HandleAmplienceError(response *http.Response) *resource.RetryError {
+	if response == nil {
+		return resource.NonRetryableError(fmt.Errorf("received nil response, unable to handle error"))
+	}
+	errResponseBody := ErrorResponse{}
 	switch response.StatusCode {
 	case http.StatusOK, http.StatusCreated, http.StatusAccepted, http.StatusNoContent:
 		return nil
 	case http.StatusInternalServerError:
-		return resource.RetryableError(fmt.Errorf("retryable error with code %d received: %s", response.StatusCode, response.Status))
+		data, err := ioutil.ReadAll(response.Body)
+		if err != nil {
+			return resource.NonRetryableError(fmt.Errorf("could not read error response body: %w", err))
+		}
+		err = json.Unmarshal(data, &errResponseBody)
+		if err != nil {
+			return resource.NonRetryableError(fmt.Errorf("could not unmarshal error response body: %w", err))
+		}
+		return resource.RetryableError(fmt.Errorf("retryable error with code %d received: %s\n "+
+			"Amplience Error Response: %s", response.StatusCode, response.Status, StringFormatObject(errResponseBody)))
 	default:
-		return resource.NonRetryableError(fmt.Errorf("non retryable error with code %d received: %s", response.StatusCode, response.Status))
+		data, err := ioutil.ReadAll(response.Body)
+		if err != nil {
+			return resource.NonRetryableError(fmt.Errorf("could not read error response body: %w", err))
+		}
+		err = json.Unmarshal(data, &errResponseBody)
+		if err != nil {
+			return resource.NonRetryableError(fmt.Errorf("could not unmarshal error response body: %w", err))
+		}
+		return resource.NonRetryableError(fmt.Errorf("non retryable error with code %d received: %s\n "+
+			"Amplience Error Response: %s", response.StatusCode, response.Status, StringFormatObject(errResponseBody)))
 	}
 }
 
@@ -126,4 +148,17 @@ type authResponseBody struct {
 	AccessToken      string `json:"access_token"`
 	SessionExpiresIn int    `json:"session_expires_in"`
 	ExpiresIn        int    `json:"expires_in"`
+}
+
+type ErrorResponseItem struct {
+	Level        string `json:"level"`
+	Code         string `json:"code"`
+	Message      string `json:"message"`
+	Property     string `json:"property"`
+	Entity       string `json:"entity"`
+	InvalidValue string `json:"invalidValue"`
+}
+
+type ErrorResponse struct {
+	Errors []ErrorResponseItem `json:"errors"`
 }