+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+(Unreleased)
+============
+- Improved Error messages for more user friendly error responses and propagation of Amplience API error response
+- Resource Webhook: Fixed a bug where creating a Webhook without a `custom_payload` would return an error
+
 0.1.0 (2021-03-17)
 ==================
 - **New resource:** `amplience_content_repository`